+++ Program.cs
@@ -39,6 +39,7 @@ builder.Services.AddScoped<IFundingEngineService, FundingEngineService>();
 builder.Services.AddScoped<IReserveRecommendationService, ReserveRecommendationService>();
 builder.Services.AddScoped<IRecommendationApplicationService, RecommendationApplicationService>();
 builder.Services.AddScoped<IDashboardSummaryService, DashboardSummaryService>();
+builder.Services.AddScoped<IDashboardExperienceService, DashboardExperienceService>();
 builder.Services.AddScoped<IMonthlyFundingReviewService, MonthlyFundingReviewService>();
 builder.Services.AddScoped<IReserveAccountSelectionService, ReserveAccountSelectionService>();
 builder.Services.AddScoped<IReservePotActionService, ReservePotActionService>();